+++ CHANGELOG.md
@@ -1,2 +1,9 @@
+- Adds new commands for provisioning and managing Firestore databases: (#5616)
+  - firestore:databases:list
+  - firestore:databases:create
+  - firestore:databases:get
+  - firestore:databases:update
+  - firestore:databases:delete
+  - firestore:locations
 - Adds `extensions` as an option in `firebase init`.
 - Relaxed repo URI validation in ext:dev:publish (#5698).